+++ packages/editor/web/src/RichContentEditor/RichContentEditor.tsx
@@ -176,6 +176,7 @@ class RichContentEditor extends Component<RichContentEditorProps, State> {
   handleCallbacks: (newState: EditorState, biCallbacks?: BICallbacks) => void | undefined;
   contextualData: EditorContextType;
   editor: Editor & { setMode: (mode: 'render' | 'edit') => void };
+  editorWrapper: Element;
   copySource: { unregister(): void };
   updateBounds: (editorBounds?: BoundingRect) => void;
   plugins;
@@ -241,7 +242,7 @@ class RichContentEditor extends Component<RichContentEditorProps, State> {
 
   componentDidMount() {
     this.copySource = registerCopySource(this.editor);
-    preventWixFocusRingAccessibility();
+    preventWixFocusRingAccessibility(this.editorWrapper);
     this.reportDebuggingInfo();
     this.props.helpers?.onOpenEditorSuccess?.(Version.currentVersion);
   }
@@ -938,6 +939,8 @@ class RichContentEditor extends Component<RichContentEditorProps, State> {
     }
   };
 
+  setEditorWrapper = ref => ref && (this.editorWrapper = ref);
+
   render() {
     const { onError, locale, direction, showToolbars = true } = this.props;
     const { innerModal } = this.state;
@@ -969,7 +972,10 @@ class RichContentEditor extends Component<RichContentEditorProps, State> {
                 data-id={'rce'}
               >
                 {this.renderStyleTag()}
-                <div className={classNames(styles.editor, theme.editor)}>
+                <div
+                  ref={this.setEditorWrapper}
+                  className={classNames(styles.editor, theme.editor)}
+                >
                   {this.renderAccessibilityListener()}
                   {this.renderEditor()}
                   {showToolbars && this.renderToolbars()}