+++ CHANGELOG.md
@@ -16,12 +16,27 @@
     Click to see more.
   </summary>
 
+</details>
+<hr/>
+
+## 8.17.3 (Jan 19, 2021)
+### :bug: Bug Fix
+- `editor`
+  - [#2019](https://github.com/wix/ricos/pull/2019) fix inline toolbar ui bug
+## 8.17.2 (Jan 19, 2021)
+### :house: Internal
+- `common`
+  - [#2012](https://github.com/wix/ricos/pull/2012) merge translations
+
+## 8.17.0 (Jan 19, 2021)
 ### :rocket: New Feature
 - `common`
   - [#1970](https://github.com/wix-incubator/rich-content/pull/1970) new theme palette fields: `textColorLow`, `disabledColor`, `fallbackColor`
-
-</details>
-<hr/>
+### :bug: Bug Fix
+- `table`
+  - [#2002](https://github.com/wix-incubator/rich-content/pull/2002) fix reset to default colors for table
+- `table`
+  - [#1999](https://github.com/wix-incubator/rich-content/pull/1999) prevent wix focus accessibility
 
 ## 8.16.0 (Jan 18, 2021)
 ### :rocket: New Feature